backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-mail>=0.10.0",
    "flask-sqlalchemy>=3.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "lxml>=6.0.2",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pypdf>=4.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-flask>=1.3.0",
]
